Rewrite the Deno KV implementation (#1)
The core of this rewrite is that all SQLite operations now synchronously run
from a single thread, and the sqlite connection is pinned to this thread. This
cleans up the messy coordination using mutexes and spawn_blocking that
we were doing previously.

Also:

Use `chrono` for all time related things, and implement `Send` for
`Sqlite` (the DB handle).

Also implements versionstamp increment randomization - we now increment
the versionstamp by a random amount between 1 and 10, rather than always
by 1.