feat: add standalone SQLite backed denokv server (#14)
This is a standalone Deno KV server backed by SQLite. It exposes a HTTP
API that implements the KV Connect protocol and can be used from
the Deno CLI.

The server can be connected to from multiple clients at the same time,
which enables a simple way to self host a production ready Deno KV
database.